GameResults/requirements.txt
astroid==2.2.5
colorama==0.4.1
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.3.1
python-dotenv==0.10.3
selenium==3.141.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2